replace squirtgun package with amethyst
