--- amethyst-api/src/main/java/me/lucyy/watercore/api/package-info.java
@@ -1,5 +1,0 @@
-/**
- * The API used to write WaterCore modules. Please be aware that this is currently
- * under heavy development, and may change at any time.
- */
-package me.lucyy.watercore.api;
